refactor(index.ts, interface.ts): implement ILogs interface for Transaction Logs being extracted
Implement ILogs interface for type checking more ts compliant in Interface.ts

BREAKING CHANGE: Logs[], and log in GetTransactionData() function now is of type ILogs

#1
